% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.R
\name{getSetting}
\alias{getSetting}
\title{Gets the setting specified by \code{key}. The key is a string containing the
complete path to the setting, where properties are seperated using a \code{.}.}
\usage{
getSetting(key, default = NULL)
}
\arguments{
\item{key}{the settings key}

\item{default}{an optional default value}
}
\value{
the key
}
\description{
Gets the setting specified by \code{key}. The key is a string containing the
complete path to the setting, where properties are seperated using a \code{.}.
}
\examples{
\dontrun{
getSetting("shiny.host")
getSetting("timeseriesapi.filters.service")
}
}

