% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run shiny sensor web client}
\usage{
run(directory, logFile = NULL, ...)
}
\description{
\code{run} starts the Shiny application with an advanced JavaScript sensor web client
}
\examples{
\dontrun{
 run()
}
}

