% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swc.R
\name{swcI18N}
\alias{swcI18N}
\title{I18N Definition}
\usage{
swcI18N(lang, key, value)
}
\arguments{
\item{lang}{the language identifier}

\item{key}{the message key}

\item{value}{the message value}
}
\value{
a HTML \code{script} tag setting the value
}
\description{
\code{swcI18N} adds a new I18N value to the JavaScript SensorWebClient.
}
\examples{
\dontrun{
 swcI18N('eng', 'button_label', 'OK')
}
}

