% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swc.R
\name{swcIntervalInput}
\alias{swcIntervalInput}
\title{Time Interval Input}
\usage{
swcIntervalInput(id)
}
\arguments{
\item{id}{the id of the input}
}
\value{
a HTML \code{input} tag
}
\description{
\code{swcIntervalInput} adds a new reactive input that contains the
currently selected timespan for the JavaScript SensorWebClient as a
lubridate interval.
}
\examples{
\dontrun{
 swcIntervalInput('time')
}
}

