% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swcLeftPanel}
\alias{swcLeftPanel}
\title{Left Panel Definition}
\usage{
swcLeftPanel(...)
}
\arguments{
\item{...}{Unnamed items that comprise this list of tags.}
}
\value{
a HTML \code{div} tag for the left panel
}
\description{
\code{swcLeftPanel} creates the left panel of the analysis view of the
JavaScript SensorWebClient.
}
\examples{
\dontrun{
swcLeftPanel(
  plotOutput("output", width="100\%", height="100\%")
)
}
}

