% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swcRightPanel}
\alias{swcRightPanel}
\title{Right Panel Definition}
\usage{
swcRightPanel(header, ...)
}
\arguments{
\item{header}{the header of the right panel}

\item{...}{Unnamed items that comprise this list of tags.}
}
\value{
a HTML \code{div} tag for the right panel
}
\description{
\code{swcLeftPanel} creates the right panel of the analysis view of the
JavaScript SensorWebClient.
}
\examples{
\dontrun{
swcRightPanel(
  header="Parameters",
  selectInput(
    "pollutant",
    label="Pollutant",
    choices = c("NOX", "NO2", "O3", "PM10", "SO2", "CO", "PM25"),
    selected = "NOX"
  ),
  swcTimeBeginInput("begin"),
  swcTimeEndInput("end"),
  swcTimeseriesInput("series"),
)
}
}

