% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swc.R
\name{swcTimeBeginInput}
\alias{swcTimeBeginInput}
\title{Begin Time Input}
\usage{
swcTimeBeginInput(id)
}
\arguments{
\item{id}{the id of the input}
}
\value{
a HTML \code{input} tag
}
\description{
\code{swcTimeBeginInput} adds a new reactive input that contains the currently
selected start time (as a \code{POSIXct}) of the JavaScript SensorWebClient.
}
\examples{
\dontrun{
 swcTimeBeginInput('begin')
}
}

