% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swc.R
\name{swcTimeseriesInput}
\alias{swcTimeseriesInput}
\title{Timeseries Input}
\usage{
swcTimeseriesInput(id)
}
\arguments{
\item{id}{the id of the input}
}
\value{
a HTML \code{input} tag
}
\description{
`swcTimeseriesInput` adds a new reactive input that contains the currently
selected time series of the JavaScript SensorWebClient (as a
`sensorweb4R::Timeseries`. Returns a HTML `input` tag.
}
\examples{
\dontrun{
 swcTimeseriesInput('series')
}
}

