% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swcPage}
\alias{swcPage}
\title{JavaScript SensorWebClient Page Definition}
\usage{
swcPage(..., title = NULL, author = NULL, description = NULL,
  debug = FALSE)
}
\arguments{
\item{...}{Unnamed items that comprise this list of tags.}

\item{title}{the title of the page}

\item{author}{the HTML \code{meta} tag author}

\item{description}{the HTML \code{meta} tag description}

\item{debug}{indicates if the client should be run in debug mode}
}
\value{
a HTML \code{html} tag containing the client
}
\description{
\code{swcPage} creates a new page containing the JavaScript
SensorWebClient.
}
\examples{
\dontrun{
swcPage(
  swcLeftPanel(...),
  swcRightPanel(...)
)
}
}

