% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/swc.R
\name{swcTimeEndInput}
\alias{swcTimeEndInput}
\title{End Time Input}
\usage{
swcTimeEndInput(id)
}
\arguments{
\item{id}{the id of the input}
}
\value{
a HTML \code{input} tag
}
\description{
\code{swcTimeEndInput} adds a new reactive input that contains the currently
selected end time (as a \code{POSIXct}) of the JavaScript SensorWebClient.
}
\examples{
\dontrun{
 swcTimeEndInput('end')
}
}

